Restructure model
